split out testing for meta data and add the wait functor for tests using the runner
