+++ README.md
@@ -1,6 +1,6 @@
 # studio-go-runner
 
-Version: <repo-version>0.9.4-feature-165-metadata-harvest-1gRhf9</repo-version>
+Version: <repo-version>0.9.4-feature-165-metadata-harvest-1gRik0</repo-version>
 
 [![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://github.com/SentientTechnologies/studio-go-runner/blob/master/LICENSE) [![Go Report Card](https://goreportcard.com/badge/SentientTechnologies/studio-go-runner)](https://goreportcard.com/report/SentientTechnologies/studio-go-runner)[![DepShield Badge](https://depshield.sonatype.org/badges/SentientTechnologies/studio-go-runner/depshield.svg)](https://depshield.github.io)
